Add alter defaultArbitraryPropertyGenerator option
