+++ fixture-monkey/src/test/java/com/navercorp/fixturemonkey/test/FixtureMonkeyV04OptionsTest.java
@@ -1,0 +1,82 @@
+/*
+ * Fixture Monkey
+ *
+ * Copyright (c) 2021-present NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.navercorp.fixturemonkey.test;
+
+import static org.assertj.core.api.BDDAssertions.then;
+import static org.assertj.core.api.BDDAssertions.thenNoException;
+import static org.assertj.core.api.BDDAssertions.thenThrownBy;
+
+import net.jqwik.api.Property;
+
+import com.navercorp.fixturemonkey.LabMonkey;
+import com.navercorp.fixturemonkey.api.generator.ObjectArbitraryPropertyGenerator;
+import com.navercorp.fixturemonkey.resolver.RootNodeResolver;
+import com.navercorp.fixturemonkey.test.FixtureMonkeyV04TestSpecs.ComplexObject;
+
+class FixtureMonkeyV04OptionsTest {
+	@Property
+	void strictModeSetWrongExpressionThrows() {
+		LabMonkey sut = LabMonkey.labMonkeyBuilder().useExpressionStrictMode().build();
+
+		thenThrownBy(
+			() -> sut.giveMeBuilder(String.class)
+				.set("nonExistentField", 0)
+				.sample()
+		).isExactlyInstanceOf(IllegalArgumentException.class)
+			.hasMessageContaining("No matching results for given NodeResolvers.");
+	}
+
+	@Property
+	void notStrictModeSetWrongExpressionDoesNotThrows() {
+		LabMonkey sut = LabMonkey.labMonkeyBuilder().build();
+
+		thenNoException()
+			.isThrownBy(() -> sut.giveMeBuilder(String.class)
+				.set("nonExistentField", 0)
+				.sample());
+	}
+
+	@Property
+	void alterMonkeyFactory() {
+		LabMonkey sut = LabMonkey.labMonkeyBuilder()
+			.monkeyExpressionFactory((expression) -> RootNodeResolver::new)
+			.build();
+		String expected = "expected";
+
+		String actual = sut.giveMeBuilder(String.class)
+			.set("test", expected)
+			.sample();
+
+		then(actual).isEqualTo(expected);
+	}
+
+	@Property
+	void alterDefaultArbitraryPropertyGenerator() {
+		LabMonkey sut = LabMonkey.labMonkeyBuilder()
+			.defaultArbitraryPropertyGenerator(
+				(context) -> ObjectArbitraryPropertyGenerator.INSTANCE.generate(context)
+					.withNullInject(1.0)
+			)
+			.build();
+
+		ComplexObject actual = sut.giveMeOne(ComplexObject.class);
+
+		then(actual).isNull();
+	}
+}
