+++ fixture-monkey/src/test/java/com/navercorp/fixturemonkey/test/FixtureMonkeyV04Test.java
@@ -43,6 +43,8 @@ import net.jqwik.api.Property;
 import com.navercorp.fixturemonkey.ArbitraryBuilder;
 import com.navercorp.fixturemonkey.ArbitraryBuilders;
 import com.navercorp.fixturemonkey.LabMonkey;
+import com.navercorp.fixturemonkey.LabMonkeyBuilder;
+import com.navercorp.fixturemonkey.api.generator.ObjectArbitraryPropertyGenerator;
 import com.navercorp.fixturemonkey.api.type.TypeReference;
 import com.navercorp.fixturemonkey.resolver.RootNodeResolver;
 import com.navercorp.fixturemonkey.test.ComplexManipulatorTestSpecs.IntValue;
@@ -707,40 +709,6 @@ class FixtureMonkeyV04Test {
 		// then
 		String notExpected = arbitraryBuilder.sample();
 		then(actual).isNotEqualTo(notExpected);
-	}
-
-	@Property
-	void strictModeSetWrongExpressionThrows() {
-		LabMonkey labMonkey = LabMonkey.labMonkeyBuilder().useExpressionStrictMode().build();
-
-		thenThrownBy(
-			() -> labMonkey.giveMeBuilder(ComplexObject.class)
-				.set("nonExistentField", 0)
-				.sample()
-		).isExactlyInstanceOf(IllegalArgumentException.class)
-			.hasMessageContaining("No matching results for given NodeResolvers.");
-	}
-
-	@Property
-	void notStrictModeSetWrongExpressionDoesNotThrows() {
-		thenNoException()
-			.isThrownBy(() -> SUT.giveMeBuilder(ComplexObject.class)
-				.set("nonExistentField", 0)
-				.sample());
-	}
-
-	@Property
-	void alterMonkeyFactory() {
-		LabMonkey labMonkey = LabMonkey.labMonkeyBuilder()
-			.monkeyExpressionFactory((expression) -> RootNodeResolver::new)
-			.build();
-		String expected = "expected";
-
-		String actual = labMonkey.giveMeBuilder(String.class)
-			.set("test", expected)
-			.sample();
-
-		then(actual).isEqualTo(expected);
 	}
 
 	@Property
